# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "VisionFlow",
        "description": "API documentation",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "/api",
            "description": "app backend"
        }
    ],
    "paths": {
        "/project/create": {
            "post": {
                "tags": [
                    "Projects"
                ],
                "summary": "create a new project",
                "operationId": "getProjects",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "title"
                                ],
                                "properties": {
                                    "title": {
                                        "type": "string",
                                        "example": "Mein Projekt"
                                    },
                                    "project_start": {
                                        "description": "Date format: YYYY-mm-dd",
                                        "type": "string",
                                        "example": "2026-01-31"
                                    },
                                    "project_end": {
                                        "description": "Date format: YYYY-mm-dd",
                                        "type": "string",
                                        "example": "2026-01-31"
                                    },
                                    "check_period": {
                                        "description": "Date Period",
                                        "type": "App\\Global\\DatePeriod",
                                        "example": "p1d"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "204": {
                        "description": "Project created"
                    }
                }
            }
        },
        "/user/register": {
            "post": {
                "tags": [
                    "User"
                ],
                "summary": "register a new user account",
                "operationId": "registerUser",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "username, password"
                                ],
                                "properties": {
                                    "username": {
                                        "type": "string",
                                        "example": "john_doe"
                                    },
                                    "password": {
                                        "type": "string",
                                        "example": "my_pass321!"
                                    },
                                    "first_name": {
                                        "type": "string",
                                        "example": "John"
                                    },
                                    "last_name": {
                                        "type": "string",
                                        "example": "Doe"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "User registered",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "token": {
                                            "type": "string",
                                            "example": "14|dJZuxaHIoPipXlyfLQjtSAsRzYLdz8ye11q9qrTS752eb1fb"
                                        },
                                        "expires_at": {
                                            "type": "string",
                                            "example": "2026-09-03 17:07:20"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        }
    },
    "tags": [
        {
            "name": "Projects",
            "description": "Projects"
        },
        {
            "name": "User",
            "description": "User"
        }
    ]
}